Fixed to build under Mono
git-svn-id: http://npoi.googlecode.com/svn/trunk@266 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ main/DDF/EscherProperty.cs
@@ -50,8 +50,8 @@ namespace NPOI.DDF
         public EscherProperty(short propertyNumber, bool isComplex, bool isBlipId)
         {
             this.id = (short)(propertyNumber +
-                    (isComplex ? unchecked((short)0x8000) : 0x0) +
-                    (isBlipId ? (short)0x4000 : 0x0));
+                    (isComplex ? unchecked((short)0x8000) : (short)0x0) +
+                    (isBlipId ? (short)0x4000 : (short)0x0));
         }
 
         /// <summary>
